Make code compatible for 16,32 & 64 bits processors
Without this change, code not works correctly, on other processors than 16 bits. 
The reason of this problem, is time, when a unsigned int is changing from negative to positive.
In 16 bits processors, negative number is changing to positive in a next operation.
In 32/64 bits processor, negative number is changing to positive immediately.